DockerDaemonSuite: cleanup "docker0" in TearDownTest
Some tests in the `DockerDaemonSuite` make modifications to the "docker0" bridge, which affects
tests running after them if they do not delete and re-create the interface.

This may also explain the `test_init_swarm_data_path_addr ` in `docker-py` failing:

```
E   docker.errors.APIError: 400 Client Error: Bad Request ("interface eth0 has more than one IPv6 address (2001:db8:1::242:ac11:2 and fe80::42:acff:fe11:2)")
```

Which may be caused by `TestDaemonIPv6FixedCIDRAndMac `:

```
s.d.StartWithBusybox(c, "--ipv6", "--fixed-cidr-v6=2001:db8:1::/64")
```

The following tests remove the `docker0` interface:

- `DockerDaemonSuite.TestDaemonIPv6FixedCIDR`
- `DockerDaemonSuite.TestDaemonIPv6FixedCIDRAndMac`
- `DockerDaemonSuite.TestDaemonIPv6HostMode`
- `DockerDaemonSuite.TestDaemonBridgeIP`
- `DockerDaemonSuite.TestDaemonDefaultGatewayIPv4Implicit`
- `DockerDaemonSuite.TestDaemonDefaultGatewayIPv4Explicit`
- `DockerDaemonSuite.TestDaemonDefaultGatewayIPv4ExplicitOutsideContainerSubnet`
- `DockerDaemonSuite.TestDaemonDefaultNetworkInvalidClusterConfig`
- `DockerDaemonSuite.TestBridgeIPIsExcludedFromAllocatorPool`

This patch removes `docker0` after each test in the `DockerDaemonSuite`, to force
the next test to re-create it with the default configuration.
