+++ integration-cli/check_test.go
@@ -589,6 +589,10 @@ func (s *DockerDaemonSuite) TearDownTest(ctx context.Context, c *testing.T) {
 		s.d.Stop(c)
 	}
 	s.ds.TearDownTest(ctx, c)
+
+	// Some tests in this Suite make modifications to the "docker0" bridge.
+	// Remove the bridge afterwards to force the next test to re-create it
+	deleteInterface(c, "docker0")
 }
 
 func (s *DockerDaemonSuite) TearDownSuite(ctx context.Context, c *testing.T) {
