+++ integration-cli/check_windows_test.go
@@ -1,0 +1,7 @@
+package main
+
+import "testing"
+
+func createInterface(c *testing.T, ifType string, ifName string, ipNet string) {}
+
+func deleteInterface(c *testing.T, ifName string) {}
